+++ packages/webmiddle-component-cheerio-to-json/src/index.js
@@ -1,0 +1,5 @@
+import CheerioToJson from "./CheerioToJson";
+import { helpers } from "webmiddle-component-cheerio-to-virtual";
+
+export default CheerioToJson;
+export { helpers };