+++ packages/webmiddle-component-jsonselect-to-json/src/index.js
@@ -1,0 +1,5 @@
+import JSONSelectToJson from "./JSONSelectToJson";
+import { helpers } from "webmiddle-component-jsonselect-to-virtual";
+
+export default JSONSelectToJson;
+export { helpers };