+++ packages/webmiddle-service-jsonselect-to-json/src/index.js
@@ -1,5 +1,0 @@
-import JSONSelectToJson from "./JSONSelectToJson";
-import { helpers } from "webmiddle-service-jsonselect-to-virtual";
-
-export default JSONSelectToJson;
-export { helpers };